+++ gen_pids/settings.py
@@ -39,6 +39,7 @@ DMS_TITLE_EXAMPLE_ENG = "Example"
 DMS_LICENSE_SCHEME_URI = "https://spdx.org/licenses/"
 DMS_LICENSE_SCHEME_ID = "SPDX"
 DMS_LICENSE_OTHER = "LicenseRef-Other"
+DMS_LANGUAGE_SCHEME_URI = "http://lexvo.org/id/iso639-3"
 
 DMS_RELATION_TYPE_ISPARTOF = "IsPartOf"
 DMS_RELATION_TYPE_HASPART = "HasPart"